www: package all folder complexity in an esbuild plugin
