+++ .prettierignore
@@ -14,3 +14,4 @@ packages/api/src/schema/validators
 packages/api/src/schema/types.d.ts
 .env
 packages/api/src/schema/schema.yaml
+packages/www/static-build
