+++ packages/api/.prettierignore
@@ -1,1 +1,3 @@
 data
+dist
+dist-esbuild
