+++ packages/api/esbuild.mjs
@@ -1,11 +1,56 @@
-import * as esbuild from "esbuild";
-import { readFile } from "fs/promises";
+// Builds a single-file zero-dependency version of the API server
+// with an embedded static version of the frontend.
 
-let envPlugin = {
-  name: "env",
+import * as esbuild from "esbuild";
+import { readFile, readdir } from "fs/promises";
+import { resolve, relative } from "path";
+
+async function getFiles(dir) {
+  const dirents = await readdir(dir, { withFileTypes: true });
+  const files = await Promise.all(
+    dirents.map((dirent) => {
+      const res = resolve(dir, dirent.name);
+      return dirent.isDirectory() ? getFiles(res) : res;
+    })
+  );
+  return Array.prototype.concat(...files);
+}
+
+const generateFrontendManifest = async (rootPath) => {
+  const output = [`module.exports = {`];
+  let files = await getFiles(rootPath);
+  for (const file of files) {
+    const key = relative(rootPath, file);
+    output.push(`  "${key}": require("${file}"),`);
+  }
+  output.push(`};`);
+  return output.join("\n");
+};
+
+let frontendBundlePlugin = {
+  name: "frontendBundle",
   setup(build) {
-    // Any files from the static-build directory should be
-    build.onLoad({ filter: /.*static\-build.*/ }, async (args) => {
+    build.onResolve({ filter: /.*static\-build$/ }, async (args) => {
+      return {
+        path: resolve(args.resolveDir, args.path),
+        namespace: "frontendBundle",
+      };
+    });
+
+    // Any files from the static-build directory should be bundled as binary
+    build.onLoad(
+      { namespace: "frontendBundle", filter: /.*static\-build/ },
+      async (args) => {
+        return {
+          contents: await generateFrontendManifest(args.path),
+          loader: "js",
+          resolveDir: resolve(args.resolveDir, args.path),
+        };
+      }
+    );
+
+    // Any files from the static-build directory should be bundled as binary
+    build.onLoad({ filter: /.*static\-build.+/ }, async (args) => {
       const contents = await readFile(args.path);
       return {
         contents: contents,
@@ -27,9 +72,6 @@ let envPlugin = {
     },
     external: ["pg-native"],
     sourcemap: "inline",
-    // loader: {
-    //   "/home/iameli/code/studio/packages/www/static-build/index.html": "binary",
-    // },
-    plugins: [envPlugin],
+    plugins: [frontendBundlePlugin],
   });
 })();
