--- packages/www/static-manifest.js
@@ -1,233 +1,0 @@
-module.exports = {
-  "404.html": require("./static-build/404.html"),
-  "_next/static/-xX8du9EuXJHZCYd1xF5x/_buildManifest.js": require("./static-build/_next/static/-xX8du9EuXJHZCYd1xF5x/_buildManifest.js"),
-  "_next/static/-xX8du9EuXJHZCYd1xF5x/_ssgManifest.js": require("./static-build/_next/static/-xX8du9EuXJHZCYd1xF5x/_ssgManifest.js"),
-  "_next/static/chunks/1574-756089e5c089db75.js": require("./static-build/_next/static/chunks/1574-756089e5c089db75.js"),
-  "_next/static/chunks/2579-79d99e6c121bde46.js": require("./static-build/_next/static/chunks/2579-79d99e6c121bde46.js"),
-  "_next/static/chunks/3493.a6cc87f50947bf0e.js": require("./static-build/_next/static/chunks/3493.a6cc87f50947bf0e.js"),
-  "_next/static/chunks/3544-63859ab5b55a4e5e.js": require("./static-build/_next/static/chunks/3544-63859ab5b55a4e5e.js"),
-  "_next/static/chunks/3646-0e460c9365ada0da.js": require("./static-build/_next/static/chunks/3646-0e460c9365ada0da.js"),
-  "_next/static/chunks/3710-2cf21312d26019df.js": require("./static-build/_next/static/chunks/3710-2cf21312d26019df.js"),
-  "_next/static/chunks/3859-2837db2bcfb09fc6.js": require("./static-build/_next/static/chunks/3859-2837db2bcfb09fc6.js"),
-  "_next/static/chunks/3867-2a90bf2d3d9a8af5.js": require("./static-build/_next/static/chunks/3867-2a90bf2d3d9a8af5.js"),
-  "_next/static/chunks/4106-0136e1e825ce72eb.js": require("./static-build/_next/static/chunks/4106-0136e1e825ce72eb.js"),
-  "_next/static/chunks/4358-b9eb00ad14b599ed.js": require("./static-build/_next/static/chunks/4358-b9eb00ad14b599ed.js"),
-  "_next/static/chunks/4713-5af986be5e61085c.js": require("./static-build/_next/static/chunks/4713-5af986be5e61085c.js"),
-  "_next/static/chunks/5633-7467c3dfcca6366d.js": require("./static-build/_next/static/chunks/5633-7467c3dfcca6366d.js"),
-  "_next/static/chunks/5817-fa970ea3344d3099.js": require("./static-build/_next/static/chunks/5817-fa970ea3344d3099.js"),
-  "_next/static/chunks/5880-a750d4e9453c6c0f.js": require("./static-build/_next/static/chunks/5880-a750d4e9453c6c0f.js"),
-  "_next/static/chunks/594-1f7151a2f6938ea8.js": require("./static-build/_next/static/chunks/594-1f7151a2f6938ea8.js"),
-  "_next/static/chunks/6044-3d46d83ead2e0950.js": require("./static-build/_next/static/chunks/6044-3d46d83ead2e0950.js"),
-  "_next/static/chunks/629-111890b138d9239c.js": require("./static-build/_next/static/chunks/629-111890b138d9239c.js"),
-  "_next/static/chunks/6542-a27b67b6d0c8e87d.js": require("./static-build/_next/static/chunks/6542-a27b67b6d0c8e87d.js"),
-  "_next/static/chunks/6716-c7590da3747b80d1.js": require("./static-build/_next/static/chunks/6716-c7590da3747b80d1.js"),
-  "_next/static/chunks/6912-d4598afac04dd7cd.js": require("./static-build/_next/static/chunks/6912-d4598afac04dd7cd.js"),
-  "_next/static/chunks/7274-c48eb5d0cef1f3fe.js": require("./static-build/_next/static/chunks/7274-c48eb5d0cef1f3fe.js"),
-  "_next/static/chunks/7389-b832ac40815fc456.js": require("./static-build/_next/static/chunks/7389-b832ac40815fc456.js"),
-  "_next/static/chunks/8979-d2e9cd097ce0e020.js": require("./static-build/_next/static/chunks/8979-d2e9cd097ce0e020.js"),
-  "_next/static/chunks/9613-d239658885730891.js": require("./static-build/_next/static/chunks/9613-d239658885730891.js"),
-  "_next/static/chunks/9657-6ccb3c6490b152e2.js": require("./static-build/_next/static/chunks/9657-6ccb3c6490b152e2.js"),
-  "_next/static/chunks/9994-c7cd7e8168ce0be2.js": require("./static-build/_next/static/chunks/9994-c7cd7e8168ce0be2.js"),
-  "_next/static/chunks/a29ae703-2e1218f6e7533903.js": require("./static-build/_next/static/chunks/a29ae703-2e1218f6e7533903.js"),
-  "_next/static/chunks/framework-658e09affb677c4e.js": require("./static-build/_next/static/chunks/framework-658e09affb677c4e.js"),
-  "_next/static/chunks/main-aec2071e4a19959d.js": require("./static-build/_next/static/chunks/main-aec2071e4a19959d.js"),
-  "_next/static/chunks/pages/_app-6e5e0c611ef1fbaa.js": require("./static-build/_next/static/chunks/pages/_app-6e5e0c611ef1fbaa.js"),
-  "_next/static/chunks/pages/_error-a3cb1e4db1b15b71.js": require("./static-build/_next/static/chunks/pages/_error-a3cb1e4db1b15b71.js"),
-  "_next/static/chunks/pages/app/admin/keys-23e20a560be49f65.js": require("./static-build/_next/static/chunks/pages/app/admin/keys-23e20a560be49f65.js"),
-  "_next/static/chunks/pages/app/admin/objectstore-43f45b513e1ca321.js": require("./static-build/_next/static/chunks/pages/app/admin/objectstore-43f45b513e1ca321.js"),
-  "_next/static/chunks/pages/app/admin/streams-faf7cabb5e924f5b.js": require("./static-build/_next/static/chunks/pages/app/admin/streams-faf7cabb5e924f5b.js"),
-  "_next/static/chunks/pages/app/admin/tools-60632b4e99c851e3.js": require("./static-build/_next/static/chunks/pages/app/admin/tools-60632b4e99c851e3.js"),
-  "_next/static/chunks/pages/app/admin/usage-389de693d7d2d100.js": require("./static-build/_next/static/chunks/pages/app/admin/usage-389de693d7d2d100.js"),
-  "_next/static/chunks/pages/app/admin/webhooks-3cd23218f68d69e4.js": require("./static-build/_next/static/chunks/pages/app/admin/webhooks-3cd23218f68d69e4.js"),
-  "_next/static/chunks/pages/app/admin-8b83b18d12e69d2d.js": require("./static-build/_next/static/chunks/pages/app/admin-8b83b18d12e69d2d.js"),
-  "_next/static/chunks/pages/app/stream/[id]-681b20c779c064d3.js": require("./static-build/_next/static/chunks/pages/app/stream/[id]-681b20c779c064d3.js"),
-  "_next/static/chunks/pages/app/stream/new-stream-356b2263a41a2fd1.js": require("./static-build/_next/static/chunks/pages/app/stream/new-stream-356b2263a41a2fd1.js"),
-  "_next/static/chunks/pages/dashboard/assets/[id]-eb69d7fbcbf06304.js": require("./static-build/_next/static/chunks/pages/dashboard/assets/[id]-eb69d7fbcbf06304.js"),
-  "_next/static/chunks/pages/dashboard/assets-8037a345259a044c.js": require("./static-build/_next/static/chunks/pages/dashboard/assets-8037a345259a044c.js"),
-  "_next/static/chunks/pages/dashboard/billing/plans-3764eb7148dbc548.js": require("./static-build/_next/static/chunks/pages/dashboard/billing/plans-3764eb7148dbc548.js"),
-  "_next/static/chunks/pages/dashboard/billing-70c8aa3a60012553.js": require("./static-build/_next/static/chunks/pages/dashboard/billing-70c8aa3a60012553.js"),
-  "_next/static/chunks/pages/dashboard/developers/api-keys-e4f1837c74dbf06c.js": require("./static-build/_next/static/chunks/pages/dashboard/developers/api-keys-e4f1837c74dbf06c.js"),
-  "_next/static/chunks/pages/dashboard/developers/signing-keys-3f2e468f119280e5.js": require("./static-build/_next/static/chunks/pages/dashboard/developers/signing-keys-3f2e468f119280e5.js"),
-  "_next/static/chunks/pages/dashboard/developers/webhooks/[id]-253f771d3c83217c.js": require("./static-build/_next/static/chunks/pages/dashboard/developers/webhooks/[id]-253f771d3c83217c.js"),
-  "_next/static/chunks/pages/dashboard/developers/webhooks-9e1f2584bea0820f.js": require("./static-build/_next/static/chunks/pages/dashboard/developers/webhooks-9e1f2584bea0820f.js"),
-  "_next/static/chunks/pages/dashboard/sessions-b9dd1d9b309850ce.js": require("./static-build/_next/static/chunks/pages/dashboard/sessions-b9dd1d9b309850ce.js"),
-  "_next/static/chunks/pages/dashboard/stream-health-44ec3a5042f9d53a.js": require("./static-build/_next/static/chunks/pages/dashboard/stream-health-44ec3a5042f9d53a.js"),
-  "_next/static/chunks/pages/dashboard/streams/[id]-ffc60dd4203044af.js": require("./static-build/_next/static/chunks/pages/dashboard/streams/[id]-ffc60dd4203044af.js"),
-  "_next/static/chunks/pages/dashboard/streams-7ae72c2ff5db5d86.js": require("./static-build/_next/static/chunks/pages/dashboard/streams-7ae72c2ff5db5d86.js"),
-  "_next/static/chunks/pages/dashboard/usage-82785f60e526a00d.js": require("./static-build/_next/static/chunks/pages/dashboard/usage-82785f60e526a00d.js"),
-  "_next/static/chunks/pages/dashboard-8e72c8254355550a.js": require("./static-build/_next/static/chunks/pages/dashboard-8e72c8254355550a.js"),
-  "_next/static/chunks/pages/forgot-password-36e85d6a845a1adb.js": require("./static-build/_next/static/chunks/pages/forgot-password-36e85d6a845a1adb.js"),
-  "_next/static/chunks/pages/index-5f848474147d110e.js": require("./static-build/_next/static/chunks/pages/index-5f848474147d110e.js"),
-  "_next/static/chunks/pages/login-7313fdf772fce38c.js": require("./static-build/_next/static/chunks/pages/login-7313fdf772fce38c.js"),
-  "_next/static/chunks/pages/register-95771b9120e1d988.js": require("./static-build/_next/static/chunks/pages/register-95771b9120e1d988.js"),
-  "_next/static/chunks/pages/reset-password-f9d631a46177475b.js": require("./static-build/_next/static/chunks/pages/reset-password-f9d631a46177475b.js"),
-  "_next/static/chunks/pages/verify-aa22d694db8334e9.js": require("./static-build/_next/static/chunks/pages/verify-aa22d694db8334e9.js"),
-  "_next/static/chunks/polyfills-c67a75d1b6f99dc8.js": require("./static-build/_next/static/chunks/polyfills-c67a75d1b6f99dc8.js"),
-  "_next/static/chunks/webpack-b1e2b27b86526c81.js": require("./static-build/_next/static/chunks/webpack-b1e2b27b86526c81.js"),
-  "_next/static/css/4ce53f5670541113.css": require("./static-build/_next/static/css/4ce53f5670541113.css"),
-  "_next/static/css/f8f9a93d01245417.css": require("./static-build/_next/static/css/f8f9a93d01245417.css"),
-  "android-chrome-192x192.png": require("./static-build/android-chrome-192x192.png"),
-  "android-chrome-256x256.png": require("./static-build/android-chrome-256x256.png"),
-  "android-chrome-512x512.png": require("./static-build/android-chrome-512x512.png"),
-  "app/admin/keys.html": require("./static-build/app/admin/keys.html"),
-  "app/admin/objectstore.html": require("./static-build/app/admin/objectstore.html"),
-  "app/admin/streams.html": require("./static-build/app/admin/streams.html"),
-  "app/admin/tools.html": require("./static-build/app/admin/tools.html"),
-  "app/admin/usage.html": require("./static-build/app/admin/usage.html"),
-  "app/admin/webhooks.html": require("./static-build/app/admin/webhooks.html"),
-  "app/admin.html": require("./static-build/app/admin.html"),
-  "app/stream/[id].html": require("./static-build/app/stream/[id].html"),
-  "app/stream/new-stream.html": require("./static-build/app/stream/new-stream.html"),
-  "apple-touch-icon.png": require("./static-build/apple-touch-icon.png"),
-  "browserconfig.xml": require("./static-build/browserconfig.xml"),
-  "dashboard/assets/[id].html": require("./static-build/dashboard/assets/[id].html"),
-  "dashboard/assets.html": require("./static-build/dashboard/assets.html"),
-  "dashboard/billing/plans.html": require("./static-build/dashboard/billing/plans.html"),
-  "dashboard/billing.html": require("./static-build/dashboard/billing.html"),
-  "dashboard/developers/api-keys.html": require("./static-build/dashboard/developers/api-keys.html"),
-  "dashboard/developers/signing-keys.html": require("./static-build/dashboard/developers/signing-keys.html"),
-  "dashboard/developers/webhooks/[id].html": require("./static-build/dashboard/developers/webhooks/[id].html"),
-  "dashboard/developers/webhooks.html": require("./static-build/dashboard/developers/webhooks.html"),
-  "dashboard/sessions.html": require("./static-build/dashboard/sessions.html"),
-  "dashboard/stream-health.html": require("./static-build/dashboard/stream-health.html"),
-  "dashboard/streams/[id].html": require("./static-build/dashboard/streams/[id].html"),
-  "dashboard/streams.html": require("./static-build/dashboard/streams.html"),
-  "dashboard/usage.html": require("./static-build/dashboard/usage.html"),
-  "dashboard.html": require("./static-build/dashboard.html"),
-  "favicon-16x16.png": require("./static-build/favicon-16x16.png"),
-  "favicon-32x32.png": require("./static-build/favicon-32x32.png"),
-  "favicon.ico": require("./static-build/favicon.ico"),
-  "fonts/Matter/Matter-Bold.eot": require("./static-build/fonts/Matter/Matter-Bold.eot"),
-  "fonts/Matter/Matter-Bold.ttf": require("./static-build/fonts/Matter/Matter-Bold.ttf"),
-  "fonts/Matter/Matter-Bold.woff": require("./static-build/fonts/Matter/Matter-Bold.woff"),
-  "fonts/Matter/Matter-Bold.woff2": require("./static-build/fonts/Matter/Matter-Bold.woff2"),
-  "fonts/Matter/Matter-BoldItalic.eot": require("./static-build/fonts/Matter/Matter-BoldItalic.eot"),
-  "fonts/Matter/Matter-BoldItalic.ttf": require("./static-build/fonts/Matter/Matter-BoldItalic.ttf"),
-  "fonts/Matter/Matter-BoldItalic.woff": require("./static-build/fonts/Matter/Matter-BoldItalic.woff"),
-  "fonts/Matter/Matter-BoldItalic.woff2": require("./static-build/fonts/Matter/Matter-BoldItalic.woff2"),
-  "fonts/Matter/Matter-Heavy.eot": require("./static-build/fonts/Matter/Matter-Heavy.eot"),
-  "fonts/Matter/Matter-Heavy.ttf": require("./static-build/fonts/Matter/Matter-Heavy.ttf"),
-  "fonts/Matter/Matter-Heavy.woff": require("./static-build/fonts/Matter/Matter-Heavy.woff"),
-  "fonts/Matter/Matter-Heavy.woff2": require("./static-build/fonts/Matter/Matter-Heavy.woff2"),
-  "fonts/Matter/Matter-HeavyItalic.eot": require("./static-build/fonts/Matter/Matter-HeavyItalic.eot"),
-  "fonts/Matter/Matter-HeavyItalic.ttf": require("./static-build/fonts/Matter/Matter-HeavyItalic.ttf"),
-  "fonts/Matter/Matter-HeavyItalic.woff": require("./static-build/fonts/Matter/Matter-HeavyItalic.woff"),
-  "fonts/Matter/Matter-HeavyItalic.woff2": require("./static-build/fonts/Matter/Matter-HeavyItalic.woff2"),
-  "fonts/Matter/Matter-Light.eot": require("./static-build/fonts/Matter/Matter-Light.eot"),
-  "fonts/Matter/Matter-Light.ttf": require("./static-build/fonts/Matter/Matter-Light.ttf"),
-  "fonts/Matter/Matter-Light.woff": require("./static-build/fonts/Matter/Matter-Light.woff"),
-  "fonts/Matter/Matter-Light.woff2": require("./static-build/fonts/Matter/Matter-Light.woff2"),
-  "fonts/Matter/Matter-LightItalic.eot": require("./static-build/fonts/Matter/Matter-LightItalic.eot"),
-  "fonts/Matter/Matter-LightItalic.ttf": require("./static-build/fonts/Matter/Matter-LightItalic.ttf"),
-  "fonts/Matter/Matter-LightItalic.woff": require("./static-build/fonts/Matter/Matter-LightItalic.woff"),
-  "fonts/Matter/Matter-LightItalic.woff2": require("./static-build/fonts/Matter/Matter-LightItalic.woff2"),
-  "fonts/Matter/Matter-Medium.eot": require("./static-build/fonts/Matter/Matter-Medium.eot"),
-  "fonts/Matter/Matter-Medium.ttf": require("./static-build/fonts/Matter/Matter-Medium.ttf"),
-  "fonts/Matter/Matter-Medium.woff": require("./static-build/fonts/Matter/Matter-Medium.woff"),
-  "fonts/Matter/Matter-Medium.woff2": require("./static-build/fonts/Matter/Matter-Medium.woff2"),
-  "fonts/Matter/Matter-MediumItalic.eot": require("./static-build/fonts/Matter/Matter-MediumItalic.eot"),
-  "fonts/Matter/Matter-MediumItalic.ttf": require("./static-build/fonts/Matter/Matter-MediumItalic.ttf"),
-  "fonts/Matter/Matter-MediumItalic.woff": require("./static-build/fonts/Matter/Matter-MediumItalic.woff"),
-  "fonts/Matter/Matter-MediumItalic.woff2": require("./static-build/fonts/Matter/Matter-MediumItalic.woff2"),
-  "fonts/Matter/Matter-Regular.eot": require("./static-build/fonts/Matter/Matter-Regular.eot"),
-  "fonts/Matter/Matter-Regular.ttf": require("./static-build/fonts/Matter/Matter-Regular.ttf"),
-  "fonts/Matter/Matter-Regular.woff": require("./static-build/fonts/Matter/Matter-Regular.woff"),
-  "fonts/Matter/Matter-Regular.woff2": require("./static-build/fonts/Matter/Matter-Regular.woff2"),
-  "fonts/Matter/Matter-RegularItalic.eot": require("./static-build/fonts/Matter/Matter-RegularItalic.eot"),
-  "fonts/Matter/Matter-RegularItalic.ttf": require("./static-build/fonts/Matter/Matter-RegularItalic.ttf"),
-  "fonts/Matter/Matter-RegularItalic.woff": require("./static-build/fonts/Matter/Matter-RegularItalic.woff"),
-  "fonts/Matter/Matter-RegularItalic.woff2": require("./static-build/fonts/Matter/Matter-RegularItalic.woff2"),
-  "fonts/Matter/Matter-SemiBold.eot": require("./static-build/fonts/Matter/Matter-SemiBold.eot"),
-  "fonts/Matter/Matter-SemiBold.ttf": require("./static-build/fonts/Matter/Matter-SemiBold.ttf"),
-  "fonts/Matter/Matter-SemiBold.woff": require("./static-build/fonts/Matter/Matter-SemiBold.woff"),
-  "fonts/Matter/Matter-SemiBold.woff2": require("./static-build/fonts/Matter/Matter-SemiBold.woff2"),
-  "fonts/Matter/Matter-SemiBoldItalic.eot": require("./static-build/fonts/Matter/Matter-SemiBoldItalic.eot"),
-  "fonts/Matter/Matter-SemiBoldItalic.ttf": require("./static-build/fonts/Matter/Matter-SemiBoldItalic.ttf"),
-  "fonts/Matter/Matter-SemiBoldItalic.woff": require("./static-build/fonts/Matter/Matter-SemiBoldItalic.woff"),
-  "fonts/Matter/Matter-SemiBoldItalic.woff2": require("./static-build/fonts/Matter/Matter-SemiBoldItalic.woff2"),
-  "fonts/Matter/demo.html": require("./static-build/fonts/Matter/demo.html"),
-  "fonts/Matter/stylesheet.css": require("./static-build/fonts/Matter/stylesheet.css"),
-  "forgot-password.html": require("./static-build/forgot-password.html"),
-  "hero-videos/dj.mp4": require("./static-build/hero-videos/dj.mp4"),
-  "hero-videos/makeup_tutorial.mp4": require("./static-build/hero-videos/makeup_tutorial.mp4"),
-  "hero-videos/poster.png": require("./static-build/hero-videos/poster.png"),
-  "hero-videos/singer.mp4": require("./static-build/hero-videos/singer.mp4"),
-  "img/404.png": require("./static-build/img/404.png"),
-  "img/OG.png": require("./static-build/img/OG.png"),
-  "img/angled-corner.svg": require("./static-build/img/angled-corner.svg"),
-  "img/arrow-right.svg": require("./static-build/img/arrow-right.svg"),
-  "img/chevron-down.svg": require("./static-build/img/chevron-down.svg"),
-  "img/copy.svg": require("./static-build/img/copy.svg"),
-  "img/docs/api-key/api-guide1.png": require("./static-build/img/docs/api-key/api-guide1.png"),
-  "img/docs/api-key/api-guide2.png": require("./static-build/img/docs/api-key/api-guide2.png"),
-  "img/docs/api-key/api-guide3.png": require("./static-build/img/docs/api-key/api-guide3.png"),
-  "img/docs/api-key/api-guide4.png": require("./static-build/img/docs/api-key/api-guide4.png"),
-  "img/docs/equalizer.png": require("./static-build/img/docs/equalizer.png"),
-  "img/docs/flag.png": require("./static-build/img/docs/flag.png"),
-  "img/docs/globe.png": require("./static-build/img/docs/globe.png"),
-  "img/docs/key.png": require("./static-build/img/docs/key.png"),
-  "img/docs/mistserver-guide1.png": require("./static-build/img/docs/mistserver-guide1.png"),
-  "img/docs/mistserver-guide2.png": require("./static-build/img/docs/mistserver-guide2.png"),
-  "img/docs/mistserver-guide3.png": require("./static-build/img/docs/mistserver-guide3.png"),
-  "img/docs/mistserver-intro.png": require("./static-build/img/docs/mistserver-intro.png"),
-  "img/docs/puzzle.png": require("./static-build/img/docs/puzzle.png"),
-  "img/docs/secret-stream-key.png": require("./static-build/img/docs/secret-stream-key.png"),
-  "img/docs/shield.png": require("./static-build/img/docs/shield.png"),
-  "img/docs/stopwatch.png": require("./static-build/img/docs/stopwatch.png"),
-  "img/docs/thumbnail.jpg": require("./static-build/img/docs/thumbnail.jpg"),
-  "img/docs/thumbnail2.jpg": require("./static-build/img/docs/thumbnail2.jpg"),
-  "img/gradient-colorful.svg": require("./static-build/img/gradient-colorful.svg"),
-  "img/gradient-violet.svg": require("./static-build/img/gradient-violet.svg"),
-  "img/help.svg": require("./static-build/img/help.svg"),
-  "img/heros/compare.png": require("./static-build/img/heros/compare.png"),
-  "img/icons/checked.svg": require("./static-build/img/icons/checked.svg"),
-  "img/icons/curly-braces.svg": require("./static-build/img/icons/curly-braces.svg"),
-  "img/icons/database.svg": require("./static-build/img/icons/database.svg"),
-  "img/icons/documentation.svg": require("./static-build/img/icons/documentation.svg"),
-  "img/icons/horn.svg": require("./static-build/img/icons/horn.svg"),
-  "img/icons/hyperlink.svg": require("./static-build/img/icons/hyperlink.svg"),
-  "img/icons/info.svg": require("./static-build/img/icons/info.svg"),
-  "img/icons/ipfs-logo.svg": require("./static-build/img/icons/ipfs-logo.svg"),
-  "img/icons/moon.svg": require("./static-build/img/icons/moon.svg"),
-  "img/icons/polygon.svg": require("./static-build/img/icons/polygon.svg"),
-  "img/icons/polygonWithoutBorderBottom.svg": require("./static-build/img/icons/polygonWithoutBorderBottom.svg"),
-  "img/icons/question.svg": require("./static-build/img/icons/question.svg"),
-  "img/icons/sun.svg": require("./static-build/img/icons/sun.svg"),
-  "img/icons/support.svg": require("./static-build/img/icons/support.svg"),
-  "img/icons/upcoming.svg": require("./static-build/img/icons/upcoming.svg"),
-  "img/investors/collaborative-fund.svg": require("./static-build/img/investors/collaborative-fund.svg"),
-  "img/investors/compound.svg": require("./static-build/img/investors/compound.svg"),
-  "img/investors/dgc.svg": require("./static-build/img/investors/dgc.svg"),
-  "img/investors/northzone.svg": require("./static-build/img/investors/northzone.svg"),
-  "img/investors/notation.svg": require("./static-build/img/investors/notation.svg"),
-  "img/korkuma-iphone.png": require("./static-build/img/korkuma-iphone.png"),
-  "img/korkuma.jpg": require("./static-build/img/korkuma.jpg"),
-  "img/logo.svg": require("./static-build/img/logo.svg"),
-  "img/prefooter-background.svg": require("./static-build/img/prefooter-background.svg"),
-  "img/share-icon.png": require("./static-build/img/share-icon.png"),
-  "img/star.svg": require("./static-build/img/star.svg"),
-  "img/temp/example-cloud.png": require("./static-build/img/temp/example-cloud.png"),
-  "img/testimonials/cbs-interactive/avatar.png": require("./static-build/img/testimonials/cbs-interactive/avatar.png"),
-  "img/testimonials/cbs-interactive/logo.svg": require("./static-build/img/testimonials/cbs-interactive/logo.svg"),
-  "img/testimonials/houseparty/avatar.png": require("./static-build/img/testimonials/houseparty/avatar.png"),
-  "img/testimonials/houseparty/logo.svg": require("./static-build/img/testimonials/houseparty/logo.svg"),
-  "img/testimonials/playdj-tv/avatar.png": require("./static-build/img/testimonials/playdj-tv/avatar.png"),
-  "img/testimonials/playdj-tv/logo.svg": require("./static-build/img/testimonials/playdj-tv/logo.svg"),
-  "img/triangle-bg.svg": require("./static-build/img/triangle-bg.svg"),
-  "img/videoPlayer.svg": require("./static-build/img/videoPlayer.svg"),
-  "img/violet-ellipsis.svg": require("./static-build/img/violet-ellipsis.svg"),
-  "index.html": require("./static-build/index.html"),
-  "login.html": require("./static-build/login.html"),
-  "markdown.css": require("./static-build/markdown.css"),
-  "mstile-150x150.png": require("./static-build/mstile-150x150.png"),
-  "pricing/group.png": require("./static-build/pricing/group.png"),
-  "register.html": require("./static-build/register.html"),
-  "reset-password.html": require("./static-build/reset-password.html"),
-  "reset.css": require("./static-build/reset.css"),
-  "robots.txt": require("./static-build/robots.txt"),
-  "safari-pinned-tab.svg": require("./static-build/safari-pinned-tab.svg"),
-  "site.webmanifest": require("./static-build/site.webmanifest"),
-  "sitemap-0.xml": require("./static-build/sitemap-0.xml"),
-  "sitemap.xml": require("./static-build/sitemap.xml"),
-  "verify.html": require("./static-build/verify.html"),
-};
